+++ examples/simple/tests/test_users_int.rs
@@ -5,7 +5,6 @@ use anyhow::Result;
 #[cfg(test)]
 mod utils;
 
-use hyper::{body::to_bytes, Method};
 use serde_json::Value;
 use ulid::Ulid;
 use utils::TestUtils;
@@ -17,16 +16,15 @@ async fn test_get_username_success() -> Result<()> {
     let username = Ulid::new().to_string();
     let token = utils.create_jwt(&username).await?;
 
-    let req = utils
+    let resp = utils
         .http
-        .call(Method::GET, "/username", Value::Null, Some(&token))?;
-
-    let resp = utils.http_client.request(req).await?;
+        .get_json("/username", Some(&token))
+        .send()
+        .await?;
 
     let status = resp.status();
-    let body = to_bytes(resp.into_body()).await?;
 
-    let json: Value = serde_json::from_slice(&body)?;
+    let json = resp.json::<Value>().await?;
 
     assert_eq!(status, 200);
     assert_eq!(json["username"], username);