+++ nakago_axum/src/auth/jwks.rs
@@ -5,10 +5,7 @@ use axum::extract::FromRef;
 use biscuit::{
     jwk::{AlgorithmParameters, JWKSet, JWK},
     jws::Secret,
-    Empty,
 };
-use hyper::{self, body::to_bytes, client::HttpConnector, Body, Method, Request};
-use hyper_tls::HttpsConnector;
 use nakago::{self, inject, Inject, Provider, Tag};
 use nakago_derive::Provider;
 use thiserror::Error;
@@ -16,10 +13,10 @@ use thiserror::Error;
 use super::Config;
 
 /// The JWKS Tag
-pub const JWKS: Tag<JWKSet<Empty>> = Tag::new("auth::JWKS");
+pub const JWKS: Tag<JWKSet<biscuit::Empty>> = Tag::new("auth::JWKS");
 
 /// Get the default set of JWKS keys
-pub async fn init(config: Config) -> JWKSet<Empty> {
+pub async fn init(config: Config) -> JWKSet<biscuit::Empty> {
     let jwks_client = Client::new(config);
 
     jwks_client
@@ -31,39 +28,28 @@ pub async fn init(config: Config) -> JWKSet<Empty> {
 /// A struct that can retrieve `JWKSet` from a configured Auth url
 pub struct Client {
     config: Config,
-    client: hyper::Client<HttpsConnector<HttpConnector>>,
 }
 
 impl Client {
     /// Create a new instance of the `Client` with the given config Arc reference
     pub fn new(config: Config) -> Self {
-        Client {
-            client: hyper::Client::builder().build::<_, Body>(HttpsConnector::new()),
-            config,
-        }
+        Client { config }
     }
 
     /// Get a `JWKSet` from the configured Auth url
-    pub async fn get_key_set(&self) -> anyhow::Result<JWKSet<Empty>> {
-        let url = format!("{}/.well-known/jwks.json", &self.config.url);
-
-        debug!("Fetching keys from '{}'", url);
-
-        let req = Request::builder()
-            .method(Method::GET)
-            .uri(url)
-            .body(Body::empty())?;
-
-        let response = self.client.request(req).await?;
-        let body = to_bytes(response.into_body()).await?;
-        let jwks = serde_json::from_slice::<JWKSet<Empty>>(&body)?;
+    pub async fn get_key_set(&self) -> anyhow::Result<JWKSet<biscuit::Empty>> {
+        let response = reqwest::get(format!("{}/.well-known/jwks.json", &self.config.url)).await?;
+        let jwks = response.json::<JWKSet<biscuit::Empty>>().await?;
 
         Ok(jwks)
     }
 }
 
 /// A convenience function to get a particular key from a key set, and convert it into a secret
-pub fn get_secret_from_key_set(jwks: &JWKSet<Empty>, key_id: &str) -> Result<Secret, ClientError> {
+pub fn get_secret_from_key_set(
+    jwks: &JWKSet<biscuit::Empty>,
+    key_id: &str,
+) -> Result<Secret, ClientError> {
     let jwk = get_key(jwks, key_id)?;
     let secret = get_secret(jwk)?;
 
@@ -71,14 +57,17 @@ pub fn get_secret_from_key_set(jwks: &JWKSet<Empty>, key_id: &str) -> Result<Sec
 }
 
 /// Get a particular key from a key set by id
-pub fn get_key(jwks: &JWKSet<Empty>, key_id: &str) -> Result<JWK<Empty>, ClientError> {
+pub fn get_key(
+    jwks: &JWKSet<biscuit::Empty>,
+    key_id: &str,
+) -> Result<JWK<biscuit::Empty>, ClientError> {
     let key = jwks.find(key_id).ok_or(ClientError::MissingKeyId)?.clone();
 
     Ok(key)
 }
 
 /// Convert a JWK into a Secret
-pub fn get_secret(jwk: JWK<Empty>) -> Result<Secret, ClientError> {
+pub fn get_secret(jwk: JWK<biscuit::Empty>) -> Result<Secret, ClientError> {
     let secret = match jwk.algorithm {
         AlgorithmParameters::RSA(rsa_key) => rsa_key.jws_public_key_secret(),
         _ => return Err(ClientError::SecretKeyError),
@@ -101,7 +90,7 @@ pub enum ClientError {
 
 /// Provide the Json Web Key Set
 ///
-/// **Provides:** `Arc<JWKSet<Empty>>`
+/// **Provides:** `Arc<JWKSet<biscuit::Empty>>`
 ///
 /// **Depends on:**
 ///   - `<Config>` - requires that `C` fulfills the `Config: FromRef<C>` constraint
@@ -131,11 +120,11 @@ impl<C: nakago::Config> Provide<C> {
 
 #[Provider]
 #[async_trait]
-impl<C: nakago::Config> Provider<JWKSet<Empty>> for Provide<C>
+impl<C: nakago::Config> Provider<JWKSet<biscuit::Empty>> for Provide<C>
 where
     Config: FromRef<C>,
 {
-    async fn provide(self: Arc<Self>, i: Inject) -> inject::Result<Arc<JWKSet<Empty>>> {
+    async fn provide(self: Arc<Self>, i: Inject) -> inject::Result<Arc<JWKSet<biscuit::Empty>>> {
         let config = if let Some(tag) = self.config_tag {
             i.get(tag).await?
         } else {